+++ src/pages/Organization/index.tsx
@@ -13,8 +13,8 @@ import {
 } from "@heroicons/react/outline";
 import { SearchIcon } from "@heroicons/react/solid";
 import logo from "images/gotsport-lt.png";
-import { OrgSearchForm } from "components/OrgSearchForm";
-import { OrgTable } from "components/OrgTable";
+import { OrgSearchForm } from "pages/Organization/OrgSearchForm";
+import { OrgTable } from "pages/Organization/OrgTable";
 
 export type Organization = {
   name: string;
@@ -245,18 +245,18 @@ export default function Organizations() {
                 </nav>
                 <button
                   type="button"
-                  className="inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+                  className="inline-flex items-center px-4 py-2 border border-gray-300 shadow-sm text-sm font-medium rounded-md text-gray-700 bg-mono-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
                 >
                   Merge Orgs
                 </button>
               </div>
             </div>
             <div className="max-w-7xl mx-auto px-4 sm:px-6 md:px-8">
-              <div className="w-full bg-white drop-shadow-md px-4 py-4 my-4">
+              <div className="w-full bg-mono-white drop-shadow-md px-4 py-4 my-4">
                 <OrgSearchForm setOrgs={setOrgs} />
               </div>
               {/* search results here */}
-              <div className="w-full bg-white drop-shadow-md px-4 py-4 my-4">
+              <div className="w-full bg-mono-white drop-shadow-md px-4 py-4 my-4">
                 <div className="py-4 flex flex-row justify-between">
                   <span className="block">Organizations ({orgs.length})</span>
                   <button
